refactor: move database to own directory
