+++ src/database/index.ts
@@ -1,0 +1,2 @@
+export { UrlDatabase } from './url-database';
+export { UrlManager } from './url-manager';
